Switch back to using integer timestamps in the log page
The live-updating code still needs float timestamps, but
we only need them for the most recent event (so we know
the lower threshold for fetching new events). We now send
the float timestamp separately:

* in the `/log/` view, we put it in HTML content, in a <script> tag
* in the `/log_events/` view we put it in response header

The main benefit of this is smaller response sizes for the
`/log/` and `/log_events/` views.

+++ hc/front/templatetags/hc_extras.py
@@ -185,8 +185,8 @@ def now_isoformat() -> str:
 
 
 @register.filter
-def timestamp(dt: datetime) -> float:
-    return dt.timestamp()
+def timestamp(dt: datetime) -> int:
+    return int(dt.timestamp())
 
 
 @register.filter
